requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
pyngrok
python-dotenv
langchain_community